+++ Code/Mantid/Framework/Algorithms/test/TransposeTest.h
@@ -64,7 +64,7 @@ public:
     MatrixWorkspace_const_sptr inputWS = boost::dynamic_pointer_cast<MatrixWorkspace>(AnalysisDataService::Instance().retrieve("transpose_irs_r"));
 
     const size_t nHist = inputWS->getNumberHistograms();
-    const int nBins = inputWS->blocksize();
+    const size_t nBins = inputWS->blocksize();
 
     TS_ASSERT_THROWS(transpose->execute(), std::runtime_error);
     TS_ASSERT(!transpose->isExecuted());
